[refactor] type 모듈화
